Remove vendored texttable & implement bold headers
